refactor(services): rename specialized orchestrators (SLM, deployment, subagent, loop, workflow, agent) (#5040)
1. SLMDeploymentOrchestrator → SLMDeploymentBridge
   File: deployment_orchestrator.py → deployment_bridge.py (old file is compat shim)
   Call-sites: api/slm/deployments.py, deployments_api_test.py updated to import
   from deployment_bridge; initialization/lifespan.py updated

2. DeploymentOrchestrator → DeploymentCoordinator
   In deployment_bridge.py; old name re-exported via compat alias in shim

3. SubagentOrchestrator → SubagentDispatcher
   File: subagent_orchestrator.py → subagent_dispatcher.py (old file is compat shim)
   get_subagent_orchestrator alias kept; test patching updated to target dispatcher

4. AutonomousLoopOrchestrator → AutonomousLoopRunner
   Same file; get_loop_orchestrator → get_loop_runner; callers in knowledge_rag.py
   and test_autonomous_loop.py updated via import alias

5. AdvancedWorkflowOrchestrator → WorkflowCoordinator
   File: orchestrator.py → coordinator.py (old file is compat shim)
   routes.py, __init__.py updated to import from coordinator directly

6. AgentOrchestrator → DistributedAgentCoordinator
   Same file (agents/agent_orchestration/coordinator.py); get_agent_orchestrator
   alias kept in __init__.py for a2a/task_executor.py and other callers

All backward-compat shims ensure existing imports resolve without change.
All 34 orchestration tests pass.
