+++ autobot-backend/agents/__init__.py
@@ -13,6 +13,7 @@ __all__ = [
     "research_agent",
     "get_chat_agent",
     "get_rag_agent",
+    "get_distributed_agent_coordinator",
     "get_agent_orchestrator",
     "AgentType",
     "ClassificationAgent",
